requirements.txt
Django==3.2.11
djangorestframework==3.13.1
djongo==1.3.6
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
sqlparse==0.2.4
